docs,test: correct the split's stated mechanism and the redirect cost
Two documentation defects.

The gate's description was wrong about its own route. At enhanced cadence with
MDKR_SYNTH_FIELDS=1 the simulation ticks on every presented frame -- 400
[SIMHASH] rows for 400 frames, hash changing on each -- so these are not frames
that present without advancing the simulation. They are authoritative ticks
whose settled menu-idle logic draws no randomness while texture animation draws
from the presentation stream. The assertion and its control are unchanged; only
the description was overstated. Corrected in the docstring, the README section,
the census, and the lane's own output, which called them presentation-only
frames.

The census concluded "redirectable: 0" without evaluating the redirect pattern
the tree already uses. cadence_compat_rand_range keeps the authoritative stream
at the shipping cadence and uses the presentation stream only at enhanced
cadence, and every gate that records an RNG stream records the original arm:
check_authored_rng_compat.py:189 pins original, check_state_hash takes the same
default, check_weather_rng_order.py:69 pins EXPECTED_ORIGINAL_SHA256. So the 8
latent callers could be moved cadence-conditionally with no rebaseline and no
gate failure. What that spends is the second compatibility target at
math_util_native.c:296-299, which no gate holds. Recorded as the cheapest route
with the reason it was not taken, and the linear-stream argument is now scoped
to unconditional redirects, which is all it licenses.

Also: use the neighbouring u64 typedef for the draw counter, and correct the
set_rng_seed comment -- it has two callers, waves.c:745 and
online/online_race_boot.c:75, not one at a stale line number.

+++ platform/math_util_native.c
@@ -238,11 +238,13 @@ s16 gArcTanTable[1026];
  *     EXPORT(gPrevRNGSeed)
  *         .word 0x5141564D   / 'QAVM' /
  *
- * Those are the LIVE starting seeds, not placeholders: set_rng_seed() has exactly
- * one caller in the whole game (game/src/waves.c:364, `set_rng_seed('WAVF')`,
- * bracketed by save_rng_seed()/load_rng_seed()), so nothing re-seeds the
- * generator at boot and every one of a run's rand_range() draws descends from it
- * -- 98 call sites, including racer.c and particles.c.
+ * Those are the LIVE starting seeds, not placeholders. set_rng_seed() has two
+ * callers: game/src/waves.c:745 (`set_rng_seed('WAVF')`, bracketed by
+ * save_rng_seed()/load_rng_seed(), so it leaves the stream where it found it),
+ * and game/src/online/online_race_boot.c:75, which reseeds from the online
+ * launch descriptor's manifest seed and so governs only an online epoch.
+ * Offline, nothing re-seeds the generator at boot and every one of a run's
+ * rand_range() draws descends from these words.
  *
  * This file shipped 0x00051234 / 0 from the first platform commit until the
  * "closedloop" wave, invented only to make the link succeed. That put the port on
@@ -264,7 +266,7 @@ s16 gArcTanTable[1026];
 s32 gCurrentRNGSeed = DKR_RNG_SEED_ROM;
 s32 gPrevRNGSeed    = DKR_RNG_SEED_ROM;
 static u32 gPresentationRNGSeed = DKR_RNG_SEED_ROM ^ 0x50524553u;
-static unsigned long long gPresentationRNGDraws;
+static u64 gPresentationRNGDraws;
 u8  gIntDisFlag     = 0; /* EXPORT(gIntDisFlag) .byte 0x00 -- matches */
 
 /* Renderer/HUD-only randomness. It intentionally uses the ROM generator's
@@ -312,7 +314,7 @@ u32 mdkr_presentation_rng_seed(void) {
     return gPresentationRNGSeed;
 }
 
-unsigned long long mdkr_presentation_rng_draws(void) {
+u64 mdkr_presentation_rng_draws(void) {
     return gPresentationRNGDraws;
 }
