+++ platform/platform_sdl_min.c
@@ -5840,7 +5840,7 @@ static void platform_frame_sync_impl(int swap, int count_present) {
             extern int32_t gCurrentRNGSeed;
             extern int32_t gPrevRNGSeed;
             extern uint32_t mdkr_presentation_rng_seed(void);
-            extern unsigned long long mdkr_presentation_rng_draws(void);
+            extern uint64_t mdkr_presentation_rng_draws(void);
             printf("[RNGSPLIT] frame=%d sim=%08x prev=%08x pres=%08x draws=%llu\n",
                    g_frameCounter, (unsigned int)gCurrentRNGSeed,
                    (unsigned int)gPrevRNGSeed, mdkr_presentation_rng_seed(),
